MAINT: Move the `np.core.numeric` functions to their own stub file
